requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
langchain
langgraph
pydantic
groq
langchain-groq
